requirements.txt
tensorflow
numpy
pandas
PyTDC
rdkit-pypi
sklearn
